refactor: tidy file structure across cli and chat apps
Consolidate MSW predicates into their respective MSW setup files,
flatten unnecessary single-level nesting (lib/chat/, cli/, mock/,
network/), standardise test helper dirs to tests/support/, and
move page-level chat tests beside src/app/page.tsx.
