+++ apps/chat/src/app/chat.loading.test.tsx
@@ -5,8 +5,7 @@ import {
 	createControllableChatStream,
 } from "../../tests/support/chat-contract";
 import { test } from "../../tests/support/chat-page-fixture";
-import { browserWorker } from "../../tests/support/msw";
-import { withJsonBody } from "../../tests/support/msw-predicates";
+import { browserWorker, withJsonBody } from "../../tests/support/msw";
 
 describe("Chat UI - loading state", () => {
 	test("disables controls while a chat request is still streaming", async ({
