+++ apps/chat/src/components/progress-thread.tsx
@@ -5,8 +5,8 @@ import {
 	formatActivitySummary,
 	formatToolOutput,
 	formatToolSummary,
-} from "@/lib/chat/tool-formatting";
-import type { ToolCallViewModel } from "@/lib/chat/view-models";
+} from "@/lib/tool-formatting";
+import type { ToolCallViewModel } from "@/lib/view-models";
 
 /**
  * SVG tree connectors — pixel-aligned to a 12×12 grid so they
