+++ apps/chat/src/lib/contracts.ts
