+++ apps/chat/src/lib/debug-trace.ts
