+++ apps/chat/src/lib/stream-reader.ts
