+++ apps/chat/src/lib/tool-formatting.ts
