+++ apps/chat/src/lib/use-auto-scroll.ts
