+++ apps/chat/src/lib/use-chat.ts
