+++ apps/chat/src/lib/view-models.ts
