+++ apps/cli/src/auth/lib/device-flow.test.ts
@@ -6,9 +6,8 @@ import {
 	requestDeviceAuthorisation,
 } from "@/auth/lib/device-flow";
 import { AuthDeviceFlowError } from "@/errors";
-import { createUnsignedJwt } from "../../../tests/helpers/jwt";
-import { withFormBody } from "../../../tests/helpers/msw-predicates";
-import { mswServer } from "../../../tests/setup/msw";
+import { mswServer, withFormBody } from "../../../tests/setup/msw";
+import { createUnsignedJwt } from "../../../tests/support/jwt";
 
 const issuer = "http://localhost:8180/realms/meridian";
 const authDeviceUrl = `${issuer}/protocol/openid-connect/auth/device`;
