+++ apps/cli/src/auth/logout.test.ts
@@ -3,10 +3,9 @@ import { join } from "node:path";
 import { HttpResponse, http } from "msw";
 import { describe, expect, it } from "vitest";
 import { runCli } from "@/cli";
-import { withFormBody } from "../../tests/helpers/msw-predicates";
-import { createWritable } from "../../tests/helpers/streams";
-import { createTempHome } from "../../tests/helpers/temp-home";
-import { mswServer } from "../../tests/setup/msw";
+import { mswServer, withFormBody } from "../../tests/setup/msw";
+import { createWritable } from "../../tests/support/streams";
+import { createTempHome } from "../../tests/support/temp-home";
 
 describe("auth logout", () => {
 	it("revokes the session and deletes stored credentials", async () => {
