+++ apps/cli/src/command-helpers.ts
