+++ apps/cli/src/proposal-requests/create.ts
@@ -1,6 +1,6 @@
 import { requireAuthentication } from "@/auth/session";
 import { getProposalRequestSchema } from "@/catalogue/registry";
-import { type JsonOption, writeError } from "@/cli/command-helpers";
+import { type JsonOption, writeError } from "@/command-helpers";
 import { writeJson, writeLines } from "@/output";
 import type { ResolvedCliDependencies } from "@/runtime";
 import { readDataStore, writeDataStore } from "@/store/data";
