+++ apps/cli/src/results/command.ts
@@ -3,7 +3,7 @@ import {
 	addJsonOption,
 	getActionCommand,
 	getJsonMode,
-} from "@/cli/command-helpers";
+} from "@/command-helpers";
 import type { ResolvedCliDependencies } from "@/runtime";
 import { handleResultsGet, type ResultsGetOptions } from "./get";
