+++ apps/cli/tests/support/jwt.ts
