+++ apps/cli/tests/support/streams.ts
